LayoutTests/fast/parser/localname-case-strict.html

<!DOCTYPE html>
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<div>
This tests <a href="http://bugs.webkit.org/show_bug.cgi?id=14114">Bug 14114</a>.
</div>
<div>
Test Data:
<a id="lowercase" href="#">lowercase</a>
<A id="uppercase" href="#">uppercase</A>
</div>
<div id="console"></div>
<script>
shouldBe("document.getElementById('lowercase').localName", "'a'");
shouldBe("document.getElementById('uppercase').localName", "'a'");
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
